build(deps): remove 'write-good'
[write-good](https://github.com/btford/write-good) is a naive linter for English prose giving a lot of warnings  but no suggestion how to improve.

The desired writing style seems to be for technical writing and not the best fit for our more colloquial language in blog posts.

As this is not used in CI anywhere, I'm dropping this from the dependencies.

Maybe we find a linter for English texts that is better suited for our purpose.